Refactored package structure. Updated log entry to decide the validity and updated readme file
