+++ src/main/java/au/com/mantel/group/logs/service/analysis/LogAnalyser.java
@@ -1,6 +1,6 @@
 package au.com.mantel.group.logs.service.analysis;
 
-import au.com.mantel.group.logs.api.model.LogEntry;
+import au.com.mantel.group.logs.service.analysis.model.LogEntry;
 import au.com.mantel.group.logs.service.analysis.model.LogAnalysisResult;
 
 import java.util.List;
