+++ src/test/groovy/au/com/mantel/group/logs/service/LogAnalyticsServiceSpec.groovy
@@ -7,8 +7,8 @@ import reactor.core.publisher.Flux;
 import spock.lang.Specification;
 import spock.lang.Subject;
 
-import au.com.mantel.group.logs.api.model.LogEntry;
-import au.com.mantel.group.logs.api.model.LogAnalyticsResponse;
+import au.com.mantel.group.logs.service.analysis.model.LogEntry;
+import au.com.mantel.group.logs.model.LogAnalyticsResponse;
 import au.com.mantel.group.logs.service.analysis.LogAnalyser;
 import au.com.mantel.group.logs.service.analysis.model.LogAnalysisResult;
